ref(authentication) changing component directory to flows; its name; using css modules;
