+++ frontend/src/flows/Authentication/Authentication.module.css
@@ -1,0 +1,8 @@
+.input-container {
+  padding-top: 20px;
+}
+
+.input {
+  margin: 10px;
+  width: 100%;
+}
