+++ frontend/src/types/flow.ts
@@ -1,0 +1,5 @@
+import { JolocomWebServiceClient } from '@jolocom/web-service-client'
+
+export interface IFlowProps {
+  serviceAPI: JolocomWebServiceClient
+}
